Optionally ask for name on new ticket forms
For new tickets created by email and in order to keep the new tickets forms short, setting the user name to what's in front of the email-@ is very nice.

However, for those who heavily rely on new ticket forms and would like to have the real user names, this new tenant setting will come in handy:

Tick the checkbox next to "ask for name on new ticket forms" and users may enter their real name when submitting a new ticket by use of the forms.