Add controller, action, user classes to body
To restyle different parts of the application with a custom stylesheet, it's necessary to have elements to root your custom styles to.

This commit adds the two classes to the body tag:
* <controller_name>-<action_name> (e.g. "tickets-new")
* "user-signed-in" if a use has previously authenticated himself

Say, you want to include the public new tickets form into another website by use of an iframe, you can match the styling by rooting your custom styles to "body.tickets-new:not(.user-signed-in)".